requirements.txt
flask>=2.2.0
twilio>=8.0.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
watchdog>=3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pypdf>=3.15.0
langchain-community>=0.0.21
